Updating Github Plus some Piglin changes
Piglins shouldn't be having wrong weapons for the AI anymore. Added scared of blocks (Soul Torches, Soul Fire, Soul Lanterns - must not be in a Bastion). Added target wither skeleton AI to both Piglin and Piglin Brute.And have checks in place for targeting to make sure they can see the target and the target is alive.